requirements.txt
marshmallow
python-dotenv
click==8.0.3
colorama==0.4.5
Flask==2.2.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.8.1
flask_sqlalchemy
Flask-Migrate
psycopg2
uuid
datetime
pandas 
flask-jwt-extended
flask-cors